% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDemo.R
\docType{methods}
\name{addTwo,CIDemo-method}
\alias{addTwo,CIDemo-method}
\alias{addTwo}
\title{add two to the supplied integer}
\usage{
\S4method{addTwo}{CIDemo}(obj, x)
}
\arguments{
\item{obj}{An object of class CIDemo}

\item{x}{numeric}
}
\description{
add two to the supplied integer
}
