% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDemo.R
\docType{class}
\name{CIDemo-class}
\alias{CIDemo-class}
\alias{.CIDemo}
\alias{CIDemo}
\title{CIDemo}
\usage{
CIDemo(projectName)
}
\arguments{
\item{projectName}{A character string}
}
\value{
An object of the CIDemo class
}
\description{
A very simple S4 class with which to explore Travis-CI
}
