% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDemo.R
\docType{methods}
\name{getProjectName,CIDemo-method}
\alias{getProjectName,CIDemo-method}
\alias{getProjectName}
\title{get the project name - to be set by derived classes}
\usage{
\S4method{getProjectName}{CIDemo}(obj)
}
\arguments{
\item{obj}{An object of class CIDemo}
}
\description{
get the project name - to be set by derived classes
}
